requirements.txt
aioredis==1.3.1
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.4
APScheduler==3.7.0
arabic-reshaper==2.1.3
asgi-redis==1.1.0
asgiref==3.3.4
astroid==2.8.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
bleach==3.3.0
boto==2.49.0
boto3==1.17.61
botocore==1.20.61
Brotli==1.0.9
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect2==0.4.1
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2
django-admin-hstore-widget==1.1.0
django-background-tasks==1.2.5
django-cache-headers==0.4
django-cleanup==5.2.0
django-clearcache==1.2.1
django-compat==1.0.15
django-constrainedfilefield==4.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-crontab==0.7.1
django-csp==3.7
django-debug-toolbar==3.2.1
django-dropbox-storage==0.4.1
django-embed-video==1.4.0
django-environ==0.4.5
django-eventstream==4.1.0
django-filter==2.4.0
django-geoposition==0.3.0
django-getenv==1.3.2
django-grip==3.0.0
django-heroku==0.3.1
django-hstore==1.4.2
django-ipware==4.0.2
django-manifest-loader==1.0.0
django-pwa==1.0.10
django-ratelimit==3.0.1
django-referrer-policy==1.0
django-rest-swagger==2.2.0
django-sri==0.3.0
django-storages==1.11.1
django-template-forms==0.3.1
django-user-agents==0.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dnspython==2.1.0
docutils==0.16
dropbox==11.7.0
email-validator==1.1.2
feedparser==6.0.8
filelock==3.0.12
fonttools==4.26.2
fpdf2==2.5.1
future==0.18.2
google==3.0.0
google-api-core==1.26.3
google-api-python-client==2.14.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googlemaps==4.4.5
gripcontrol==4.1.0
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
imagesize==1.2.0
imgkit==1.2.2
importlib-metadata==4.10.0
incremental==21.3.0
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
keyring==23.0.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mailjet-rest==1.3.4
Markdown==3.3.6
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
msgpack-python==0.5.6
mysqlclient==2.0.3
numerize==0.12
numpy==1.21.4
oauth2client==4.1.3
openapi-codec==1.3.2
packaging==20.9
pdfkit==1.0.0
pefile==2019.4.18
Pillow==8.2.0
pkginfo==1.7.0
platformdirs==2.4.0
pluggy==0.13.1
ply==3.11
protobuf==3.15.8
psycopg2==2.8.6
pubcontrol==3.1.0
pushpad==0.9.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydyf==0.1.1
Pygments==2.8.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pytest==6.2.3
pytest-runner==5.3.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
python-docx==0.8.11
python-dotenv==0.17.1
python-http-client==3.3.2
python-magic==0.4.26
pytz==2021.1
PyVimeo==1.1.0
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==6.1
r-django-jet==2.0.9
readme-renderer==29.0
redis==3.5.3
reportlab==3.6.1
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
s3transfer==0.4.2
sendgrid==6.7.0
service-identity==18.1.0
sgmllib3k==1.0.0
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stone==3.2.1
timeloop==1.0.2
tinycss2==1.1.0
tinydb==4.4.0
toml==0.10.2
tqdm==4.60.0
tuspy==0.2.4
twilio==6.59.0
twine==3.4.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
ua-parser==0.10.0
unicorn==2.0.0
uritemplate==3.0.1
urllib3==1.26.4
user-agent==0.1.10
user-agents==2.2.0
uvicorn==0.14.0
virtualenv==20.4.7
weasyprint==53.2
webencodings==0.5.1
Werkzeug==0.16.1
whitenoise==5.2.0
wkhtmltopdf
wkhtmltopdf-pack-ng==0.12.3.0
wrapt==1.13.3
xhtml2pdf==0.2.5
zipp==3.4.1
zope.interface==5.4.0
zopfli==0.1.8